run_mpitests: split utilities into multi_tests.c
It is cleaner to split the utilities for multi-tests into its own source
file. Since it will only be used in run_mpitests.c, it is simpler to
just include the file as static code.
